requirements.txt
Flask==0.10.1
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-telegram-bot==3.0.0
PyYAML==3.11
Werkzeug==0.11.2
wheel==0.24.0
